[base vm] explicitly cast size_t to int

+++ src/platform/backends/shared/base_virtual_machine.h
@@ -56,7 +56,7 @@ public:
     SnapshotVista view_snapshots() const noexcept override;
     inline int get_num_snapshots() const noexcept override
     {
-        return snapshots.size();
+        return static_cast<int>(snapshots.size());
     }
     std::shared_ptr<const Snapshot> get_snapshot(const std::string& name) const override;
